feat!: rename `WebWebpackPlugin` to `WebEncodePlugin`
