+++ .changeset/eighty-snails-crash.md
@@ -1,0 +1,7 @@
+---
+"@lynx-js/template-webpack-plugin": patch
+---
+
+Add `WebEncodePlugin`.
+
+This is previously known as `WebWebpackPlugin` from `@lynx-js/web-webpack-plugin`.
