+++ .changeset/forty-garlics-build.md
@@ -2,7 +2,6 @@
 "@lynx-js/css-extract-webpack-plugin": patch
 "@lynx-js/react-refresh-webpack-plugin": patch
 "@lynx-js/react-webpack-plugin": patch
-"@lynx-js/web-webpack-plugin": patch
 ---
 
 Support `@lynx-js/template-webpack-plugin` v0.7.0.
