+++ .changeset/solid-ears-taste.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/react-rsbuild-plugin": patch
+---
+
+Use `WebEncodePlugin` instead of `WebWebpackPlugin`.
