+++ packages/webpack/template-webpack-plugin/src/web/StyleInfo.ts
@@ -2,7 +2,6 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-// LICENSE file in the root directory of this source tree.
 export interface CSSRule {
   sel: string[][][];
   decl: [string, string][];
