+++ packages/webpack/template-webpack-plugin/src/web/genStyleInfo.ts
@@ -2,9 +2,8 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import { CSS } from '@lynx-js/template-webpack-plugin';
-
-import type { CSSRule, OneInfo, StyleInfo } from '../type/StyleInfo.js';
+import type { CSSRule, OneInfo, StyleInfo } from './StyleInfo.js';
+import * as CSS from '../css/index.js';
 
 export function genStyleInfo(
   cssMap: Record<string, CSS.LynxStyleNode[]>,
