+++ packages/webpack/template-webpack-plugin/test/cases/web/all-in-one/test.config.cjs
