+++ packages/webpack/template-webpack-plugin/test/cases/web/default-export/b.js
