--- packages/webpack/web-webpack-plugin/api-extractor.json
@@ -1,7 +1,0 @@
-/**
- * Config file for API Extractor.  For more info, please visit: https://api-extractor.com
- */
-{
-  "extends": "../../../api-extractor.json",
-  "mainEntryPointFilePath": "dist/index.d.ts",
-}
