--- packages/webpack/web-webpack-plugin/etc/web-webpack-plugin.api.md
@@ -1,39 +1,0 @@
-## API Report File for "@lynx-js/web-webpack-plugin"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import type { Compilation } from 'webpack';
-import type { Compiler } from 'webpack';
-
-// Warning: (ae-missing-release-tag) "isDebug" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export function isDebug(): boolean;
-
-// Warning: (ae-missing-release-tag) "isRsdoctor" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export function isRsdoctor(): boolean;
-
-// Warning: (ae-missing-release-tag) "WebWebpackPlugin" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export class WebWebpackPlugin {
-    // (undocumented)
-    apply(compiler: Compiler): void;
-    // (undocumented)
-    static BEFORE_ENCODE_HOOK_STAGE: number;
-    deleteDebuggingAssets(compilation: Compilation, assets: ({
-        name: string;
-    } | undefined)[]): void;
-    // (undocumented)
-    static ENCODE_HOOK_STAGE: number;
-    // (undocumented)
-    static name: string;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
