--- packages/webpack/web-webpack-plugin/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@lynx-js/web-webpack-plugin",
-  "version": "0.6.7",
-  "private": false,
-  "description": "TBD",
-  "keywords": [],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lynx-family/lynx-stack.git",
-    "directory": "packages/webpack/web-webpack-plugin"
-  },
-  "license": "Apache-2.0",
-  "type": "module",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "!dist/**/*.js.map",
-    "LICENSE.txt",
-    "Notice.txt",
-    "CHANGELOG.md",
-    "README.md"
-  ],
-  "scripts": {
-    "api-extractor": "api-extractor run --verbose",
-    "test": "pnpm vitest --project webpack/web-webpack-plugin"
-  },
-  "dependencies": {
-    "@lynx-js/web-style-transformer": "workspace:*"
-  },
-  "devDependencies": {
-    "@lynx-js/template-webpack-plugin": "workspace:*",
-    "@lynx-js/test-tools": "workspace:*",
-    "webpack": "^5.99.9"
-  },
-  "peerDependencies": {
-    "@lynx-js/template-webpack-plugin": "^0.6.0 || ^0.7.0"
-  }
-}
