--- packages/webpack/web-webpack-plugin/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "baseUrl": "./",
-    "rootDir": "./src",
-    "outDir": "./dist",
-  },
-  "include": ["src"],
-  "references": [
-    { "path": "../../web-platform/web-style-transformer" },
-  ],
-}
