+++ website/package.json
@@ -29,7 +29,6 @@
     "@lynx-js/rspeedy": "workspace:*",
     "@lynx-js/runtime-wrapper-webpack-plugin": "workspace:*",
     "@lynx-js/template-webpack-plugin": "workspace:*",
-    "@lynx-js/web-webpack-plugin": "workspace:*",
     "@lynx-js/webpack-runtime-globals": "workspace:*",
     "@microsoft/api-documenter": "~7.23.38",
     "@microsoft/api-extractor-model": "^7.30.6",
